pyo3-build-config: check python version and architecture in PyO3 crate
